Produce per-shepherd highlighted sheets in one go
This introduces a new option `--shepherds-combined` which controls
whether or not the highlights for the chosen shepherds appear
individually or combined onto a single page. While the default was
for the highlights to appear together on a single page, we change
the default to be separate pages per shepherd as this option turns
out to be more useful in practise.

Fixes https://github.com/PeterJCLaw/srcomp-cli/issues/2