Ensure we do actually fetch a deployed revision when asked
Previously we relied on HEAD at the remote end being a commit which
contained the deployed state, which wasn't neccesarily true. By
requesting the given commit directly we can be more sure that it'll
be downloaded. Of course this is more risky as the remote may error
about not having the object, though that should never actually
happen for a deployed commit.

This should help with https://github.com/PeterJCLaw/srcomp-cli/issues/11
but won't fully fix it.